refactor: move batch optimization functions
